+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add possible to replace target file if exists 
+### Fixed
+- Fix validator condition
 
 ## [0.3.2] – 2022-09-20
 ### Fixed